requirements.txt
asgiref==3.2.10
Django==3.1
pytz==2022.1
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0 
pyjwt==2.4.0